struct/remove-types: remove phono.BufferSize and fix examples

+++ test/test.go
@@ -16,7 +16,7 @@ var (
 		Wav1        string
 		Wav2        string
 		Mp3         string
-		Wav1Samples int64
+		Wav1Samples int
 	}{
 		Wav1:        resolvePath(testdata + "sample1.wav"), // Wav1 is the wav file with bass slap sample.
 		Wav2:        resolvePath(testdata + "sample2.wav"), // Wav2 is the wav file with trimmed reversed bass slap sample.
